requirements.txt
torch>=2.0
numpy>=1.24
gymnasium>=0.29
Pillow>=9.0
vizdoom>=1.2
